Fix misspelled icon path
The icon read note-sticky and not sticky-note and in addition the word
fontawesome was missing an e.

To check the icons, you can use the icons and emoji browser in the
documentation: https://squidfunk.github.io/mkdocs-material/reference/icons-emojis/

The search bar let's you any icon or emoji and presents you the markdown
tag. To get a filename out of it, just replace the first two dashes with
slashes:

:fontawesome-solid-note-sticky: becomes fontawesome/solid/note-sticky

Fixes #53